Rename schemes.py to schemas.py
