+++ homeassistant/components/lcn/schemas.py
